Log file and count variable
Created a system for saving all past prompts to a log file that is also viewable in the UI, and the ability to clear it with a button. (Press button twice to update UI).
Also added "count" variable that allows to generate hundreds at once and just look through the log.